Extract add component to its own file
